+++ ElvUI_EltreumUI/Localization/ptBR.lua
@@ -1599,3 +1599,5 @@ L["The text displayed when in a quest instance"] = "O texto exibido em uma inst√
 L["Thin Aura Bars"] = "Barras de Auras Finas"
 L["Exported Gradient Data:"] = true
 L["Import Gradient Data"] = true
+L["Reinstall Thin Layout"] = true
+L["Reset layout to Eltruism Thin"] = true